状态同步运行成功

+++ Scripts/Hotfix/Server/Map/FiberInit_Map.cs
@@ -2,7 +2,7 @@
 
 namespace ET.Server
 {
-    [Invoke((long)SceneType.Map)]
+    [Invoke(SceneType.Map)]
     public class FiberInit_Map: AInvokeHandler<FiberInit, ETTask>
     {
         public override async ETTask Handle(FiberInit fiberInit)
